+++ packages/style-system-demo/cypress/integration/GridColSamples.spec.ts
@@ -3,6 +3,7 @@ import {clickEl, getEl} from '../utils/utils'
 describe('Grid Col samples', () => {
   it('Responsive (desktop)', () => {
     clickEl('/grid-with-col/responsive')
+    cy.viewport(1600, 1200)
     getEl('col1').then(($col1) => {
       getEl('col2').then(($col2) => {
         assert.equal($col1.position().top, $col2.position().top, 'Col1 and col2 have to be in the same row.')
@@ -38,10 +39,10 @@ describe('Grid Col samples', () => {
     })
     getEl('col4').then(($col4) => {
       getEl('col5').then(($col5) => {
-        assert.equal($col4.position().top, $col5.position().top, 'Col1 and col5 have to be in the same row.')
+        assert.equal($col4.position().top, $col5.position().top, 'Col4 and col5 have to be in the same row.')
       })
       getEl('col6').then(($col6) => {
-        assert.equal($col4.position().top, $col6.position().top, 'Col1 and col6 have to be in the same row.')
+        assert.equal($col4.position().top, $col6.position().top, 'Col4 and col6 have to be in the same row.')
       })
     })
   })