+++ packages/style-system-demo/src/grid/col/ResponsiveGridColCode.ts
@@ -1,0 +1,12 @@
+/* tslint:disable */
+export default `<Grid spacing="0.5rem">
+  <ColRow>
+    <Col size={[6, 4, 2]} testId="col1"><OrangeBox/></Col>
+    <Col size={[6, 4, 2]} testId="col2"><YellowBox/></Col>
+    <Col size={[6, 4, 2]} testId="col3"><OrangeBox/></Col>
+    <Col size={[6, 4, 2]} testId="col4"><YellowBox/></Col>
+    <Col size={[6, 4, 2]} testId="col5"><OrangeBox/></Col>
+    <Col size={[6, 4, 2]} testId="col6"><YellowBox/></Col>
+  </ColRow>
+</Grid>
+`